LayoutTests/imported/w3c/web-platform-tests/clipboard-apis/async-unsanitized-html-formats-write-read.tentative.https-expected.txt

Body needed for test_driver.click()

FAIL Verify write and read unsanitized content to the clipboard given text/html format as input assert_equals: expected "<head> </head> <body> <p style=\"color: blue; font-size: medium; font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; white-space: normal; text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;\">Hello World</p> </body>" but got "<head> </head> <body> <p style=\"color: blue; font-style: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: auto; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: auto; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration: none;\">Hello World</p> </body>"
